% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_dates}
\alias{generate_dates}
\title{generate dates}
\usage{
generate_dates(start_date, end_date = start_date + 31, fasting = "week")
}
\arguments{
\item{start_date}{the start date}

\item{end_date}{the end date}

\item{fasting}{which days are fasting days?}
}
\value{
data.frame
}
\description{
creates a data.frame that can be easily imported into a google sheet for
tracking weight over a date range.
}
