% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_weight}
\alias{plot_weight}
\title{plot weight}
\usage{
plot_weight(weight_data, ylim = range(weight_data$weight, na.rm = TRUE))
}
\arguments{
\item{weight_data}{data.frame containing "date", "weight", "status"}

\item{ylim}{the limits in the y-direction (defaults to minimum and maximum of data)}
}
\value{
plot
}
\description{
generates a nice plot of weight over time to help keep me motivated
}
