% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_sheet_data}
\alias{read_sheet_data}
\title{read data}
\usage{
read_sheet_data(file = NULL, url = NULL)
}
\arguments{
\item{file}{a file to read a url from}

\item{url}{the actual url}
}
\value{
data.frame
}
\description{
reads the data from the existing sheet
}
\examples{
# from a file
url_file <- system.file("extdata", "gs_url", package = "trackWeight")
read_sheet_data(file = url_file)

# from the url directly
read_sheet_data(url = "https://docs.google.com/spreadsheets/d/1f7X_60NeRRfXxhc3KGO62NdIz2wwfWIrY_vpOMy9hPY/export?format=csv")

}
